% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_sin.R
\name{fn_cos}
\alias{fn_cos}
\title{fn_cos}
\usage{
fn_cos(x, k)
}
\arguments{
\item{x}{a single number. The number to be transformed}

\item{k}{a positive integer. The number of terms to be used in the series expansion beyond the constant 1.}
}
\value{
return approximation
}
\description{
computes the approximation of cos
}
\details{
This function use the infinite series expansion of the cos form to calculate cos.
}
\examples{
fn_cos(pi, 5)


}
