% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_sin.R
\name{fun_sin}
\alias{fun_sin}
\title{fun_sin}
\usage{
fun_sin(x, k)
}
\arguments{
\item{x}{a single number. The number to be transformed}

\item{k}{a positive integer. The number of terms to be used in the series expansion beyond the constant 1.}
}
\value{
return approximation of sin
}
\description{
computes the approximation of sin
}
\details{
This function use the infinite series expansion of the sin form to calculate sin.
}
\examples{
fun_sin(pi/2,5)

}
