% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_class.R
\name{make_ci_class}
\alias{make_ci_class}
\title{make_ci_class}
\usage{
make_ci_class(obj)
}
\arguments{
\item{obs}{a set of number.}
}
\value{
return ci_class object
}
\description{
create a ci_class object
}
\details{
This a constructor function for ci_class
}
\examples{
x=rnorm(100)
obj=make_ci_class(x)
obj$obs_mean
obj$obs_sd
}
