% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_class.R
\name{print.ci_class}
\alias{print.ci_class}
\title{print method for ci_class}
\usage{
\method{print}{ci_class}(x)
}
\arguments{
\item{x}{a ci_class object}
}
\value{
return a message with name of the class and the the number of observations in the ci_class object
}
\description{
print method for ci_class
}
\details{
This print method for ci_class
}
\examples{
x=rnorm(100)
obj=make_ci_class(x)
print(obj)

}
